refactor(test): split rendering style assertions
